Factor out context actions/assertions
